set review delete part

+++ controllers/reviewController.js
@@ -23,6 +23,7 @@ export function addReview(req,res){
 }
 
 export function getReviews(req,res){
+
     const user = req.user
 
     if(user == null || user.role != "admin"){
@@ -41,9 +42,39 @@ export function getReviews(req,res){
 export function deleteReview(req,res){
     const email = req.params.email
 
-    Review.deleteOne({email:email}).then(()=>{
+    if(req.user = null){
+        res.status(401).json({message:"Please login and try again"})
+        return
+    }
+
+    if(req.user.role == "admin"){
+      
+    
+
+   Review.deleteOne({email:email}).then(()=>{
         res.json({message : "Review deleted successfully"})
     }).catch(()=>{
-        res.status(500).json({error : "Review deletion"})
+        res.status(500).json({error : "Review deletion failed"})
     })
+
+    return
+
+    }
+
+    if(req.user.role == "customer"){
+
+        if(req.user.email == email){
+            Review.deleteOne({email:email}).then(()=>{
+                res.json({message : "Review deleted successfully"})
+            }).catch(()=>{
+                res.status(500).json({error : "Review deletion failed"})
+            })
+        }else{
+            res.status(403).json({message : "You are not authorized to perform this action"})
+        }
+    }
+    
 }
+
+
+
